+++ httpie/__init__.py
@@ -3,7 +3,7 @@ HTTPie: modern, user-friendly command-line HTTP client for the API era.
 
 """
 
-__version__ = '3.2.3'
-__date__ = '2024-07-10'
+__version__ = '3.2.4'
+__date__ = '2024-11-01'
 __author__ = 'Jakub Roztocil'
 __licence__ = 'BSD'