+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 This document records all notable changes to [HTTPie](https://httpie.io).
 This project adheres to [Semantic Versioning](https://semver.org/).
 
+## [3.2.4](https://github.com/httpie/cli/compare/3.2.3...3.2.4) (2024-11-01)
+
+- Fix default certs loading and unpin `requests`. ([#1596](https://github.com/httpie/cli/issues/1596))
+
 ## [3.2.3](https://github.com/httpie/cli/compare/3.2.2...3.2.3) (2024-07-10)
 
 - Fix SSL connections by pinning the `requests` version to `2.31.0`. (#1583, #1581)